--- src/Trash.js
@@ -1,26 +1,0 @@
-// import {Box, Card, CardBody, CardFooter, CardHeader, Center, Divider, Heading, Spacer} from "@chakra-ui/react";
-// import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
-// import {faRecordVinyl} from "@fortawesome/free-solid-svg-icons";
-// import React from "react";
-//
-// <>
-//   <Box mt={30}>
-//     <Center gap={5}>
-//       <Card w="sm">
-//         <CardHeader _hover={{cursor: "pointer"}}>
-//           <Image src="https://cdn.dribbble.com/users/5783048/screenshots/13902636/skull_doodle_4x.jpg"/>
-//         </CardHeader>
-//         <CardBody>
-//           <Heading size="md">1</Heading>
-//         </CardBody>
-//         <Divider color="gray"/>
-//         <CardFooter>
-//           <FontAwesomeIcon color="gray" icon={faRecordVinyl}/>
-//           <Text>$곡</Text>
-//           <Spacer/>
-//           <Text>생성일 : </Text>
-//         </CardFooter>
-//       </Card>
-//     </Center>
-//   </Box>
-// </>
