+++ src/index.js
@@ -1,3 +1,5 @@
+import React from "react";
+
 import App from "./App";
 import { ChakraProvider } from "@chakra-ui/react";
 import { createRoot } from "react-dom/client";
