--- src/page/component/NavBar.js
@@ -1,28 +1,0 @@
-import { faUserPlus } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { Button } from "@chakra-ui/react";
-import { LoginContext } from "./LogInProvider";
-import { useContext } from "react";
-import { useNavigate } from "react-router-dom";
-
-export function NavBar() {
-  const { isAuthenticated } = useContext(LoginContext);
-
-  const navigate = useNavigate();
-
-  return (
-    <>
-      {isAuthenticated() || (
-        <Button
-          borderRadius={0}
-          variant="ghost"
-          size="lg"
-          leftIcon={<FontAwesomeIcon icon={faUserPlus} />}
-          onClick={() => navigate("/signup")}
-        >
-          signup
-        </Button>
-      )}
-    </>
-  );
-}
